refactor cvs1 menu attribs into their own file
